rest project

+++ src/main/java/spring_demo/service/AccountService.java
@@ -1,5 +1,5 @@
 package spring_demo.service;
 
 public interface AccountService {
-    void transfer(long accountId, long otherAccountId, double balance);
+    void transfer(int accountId, int otherAccountId, double balance);
 }
